Hn1

+++ include/domain/entities/Staff.hpp
@@ -9,6 +9,6 @@ private:
 public:
     Staff() : Account(), salary(0.0) {}
     Staff(const string &Id, const string &Name, const string &Pass, const string &Gen, const float &Salary) : Account(Id, Name, Pass, Gen), salary(Salary) {}
-    double getSalary() const { return salary; };              // getter
-    void setSalary(double salary) { this->salary = salary; }; // setter
+    float getSalary() const { return salary; };              // getter
+    void setSalary(float salary) { this->salary = salary; }; // setter
 };
